feat: base mint setup done
