+++ src/index.tsx
@@ -1,13 +1,14 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import "./styles/index.css";
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './styles/index.css';
 
-import Landing from "./pages/Landing.jsx";
-import { Router } from "@reach/router";
-import reportWebVitals from "./reportWebVitals";
-import NFT from "./pages/NFT";
-import Roadmap from "./pages/Roadmap";
-import Tokenomics from "./pages/Tokenomics";
+import Landing from './pages/Landing.jsx';
+import { Router } from '@reach/router';
+import reportWebVitals from './reportWebVitals';
+import NFT from './pages/NFT';
+import Roadmap from './pages/Roadmap';
+import Tokenomics from './pages/Tokenomics';
+import Mint from './App';
 
 ReactDOM.render(
   <React.StrictMode>
@@ -18,10 +19,11 @@ ReactDOM.render(
         <NFT path="/nft" />
         <Roadmap path="/roadmap" />
         <Tokenomics path="/tokenomics" />
+        <Mint path="mint" />
       </Router>
     </div>
   </React.StrictMode>,
-  document.getElementById("root")
+  document.getElementById('root')
 );
 
 // If you want to start measuring performance in your app, pass a function
